+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 ## 3.1.0
 - [BREAKING CHANGE] `BlockMovementCollision` big improvements.
-  Update `onBlockedMovement` method, adds `CollisionData(normal,depth,direction,intersectionPoints)`
-  Adds `onBlockMovementUpdateVelocity` method to do override if necessary.
+  - Update `onBlockedMovement` method, adds `CollisionData(normal,depth,direction,intersectionPoints)`
+  - Adds `onBlockMovementUpdateVelocity` method to do override if necessary.
+  - Adds `setupBlockMovementCollision({bool? enabled, bool? isRigid})`
 - `HandleForces` improvements.
 - Fix typo in params that contained `intencity` renaming it to `intensity`.
 - Fix bug `playOnce` and `playOnceOther` when call again without finish the last.